minor clean ups
