+++ examples/03-company-employees/CompanyEmployeesExample.java
@@ -12,18 +12,18 @@ public class CompanyEmployeesExample {
             // Load DSL from external JSON file for better readability
             Generation result = DslDataGenerator.create()
                     .withSeed(98765L)
-                    .fromFile(new java.io.File("dsl.json"))
+                    .fromFile(new File("dsl.json"))
                     .generate();
 
             System.out.println("=== Generated JSON ===");
             System.out.println(result.asJson());
-            
+
             System.out.println("\n=== Generated SQL ===");
             result.asSqlInserts().forEach((table, sql) -> {
                 System.out.println("-- Table: " + table);
                 System.out.println(sql);
             });
-            
+
         } catch (Exception e) {
             e.printStackTrace();
         }
