+++ examples/07-custom-generator/CustomGeneratorExample.java
@@ -14,20 +14,20 @@ import java.util.Random;
  * Shows how to extend the library with domain-specific generators.
  */
 public class CustomGeneratorExample {
-    
+
     public static void main(String[] args) {
         try {
             // Initialize dependencies
             ObjectMapper mapper = new ObjectMapper();
             Faker faker = new Faker(new Random(42));
-            
+
             // Create custom employee ID generator
             Generator employeeIdGenerator = (options) -> {
                 String prefix = options.has("prefix") ? options.get("prefix").asText() : "EMP";
                 int number = faker.number().numberBetween(1000, 9999);
                 return mapper.valueToTree(prefix + "-" + String.format("%04d", number));
             };
-            
+
             // Create custom department code generator
             Generator departmentCodeGenerator = (options) -> {
                 if (options.has("department")) {
@@ -55,7 +55,7 @@ public class CustomGeneratorExample {
                 }
                 return mapper.valueToTree("UNK");
             };
-            
+
             // Create complex job level info generator that creates complete objects
             Generator jobLevelInfoGenerator = (options) -> {
                 String[] levels = {"Junior", "Mid", "Senior", "Lead"};
@@ -65,12 +65,12 @@ public class CustomGeneratorExample {
                 String[] descriptions = {
                     "Entry-level position with mentorship and learning opportunities",
                     "Experienced professional with independent project ownership",
-                    "Senior contributor with technical leadership responsibilities", 
+                    "Senior contributor with technical leadership responsibilities",
                     "Strategic technical leader driving architectural decisions"
                 };
-                
+
                 int index = faker.number().numberBetween(0, levels.length);
-                
+
                 // Create a complete job level object that will be spread using "..."
                 var jobLevel = mapper.createObjectNode();
                 jobLevel.put("level", levels[index]);
@@ -80,39 +80,39 @@ public class CustomGeneratorExample {
                 jobLevel.put("description", descriptions[index]);
                 jobLevel.put("years_experience_min", (index + 1) * 2);
                 jobLevel.put("years_experience_max", (index + 1) * 4);
-                
+
                 return jobLevel;
             };
-            
+
             System.out.println("=== Custom Generator Example ===");
             System.out.println("Generating employee data with custom business logic...");
-            
+
             // Generate data using custom generators
             Generation result = DslDataGenerator.create()
                     .withSeed(12345L)
                     .withCustomGenerator("employeeId", employeeIdGenerator)
                     .withCustomGenerator("departmentCode", departmentCodeGenerator)
                     .withCustomGenerator("jobLevelInfo", jobLevelInfoGenerator)
-                    .fromFile(new java.io.File("dsl.json"))
+                    .fromFile(new File("dsl.json"))
                     .generate();
 
             System.out.println("\n=== Generated JSON ===");
             System.out.println(result.asJson());
-            
+
             System.out.println("\n=== Generated SQL ===");
             result.asSqlInserts().forEach((table, sql) -> {
                 System.out.println("-- Table: " + table);
                 System.out.println(sql);
                 System.out.println();
             });
-            
+
             System.out.println("=== Custom Generator Features Demonstrated ===");
             System.out.println("✓ Employee IDs follow company format (EMP-XXXX)");
             System.out.println("✓ Department codes are business-appropriate abbreviations");
             System.out.println("✓ Job level info generator creates complete objects with spread operator");
             System.out.println("✓ Complex generators can return multiple fields at once");
             System.out.println("✓ Custom generators integrate seamlessly with built-in ones");
-            
+
         } catch (Exception e) {
             System.err.println("Error running custom generator example: " + e.getMessage());
             e.printStackTrace();
