+++ src/main/java/com/github/eddranca/datagenerator/EagerGeneration.java
@@ -2,11 +2,8 @@ package com.github.eddranca.datagenerator;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Stream;
 
 public class EagerGeneration extends AbstractGeneration<JsonNode> {
 
@@ -20,121 +17,5 @@ public class EagerGeneration extends AbstractGeneration<JsonNode> {
         return item;
     }
 
-    /**
-     * Fluent builder for generation operations.
-     */
-    public static class Builder {
-        private final DslDataGenerator generator;
-        private final File file;
-        private final String jsonString;
-        private final JsonNode jsonNode;
 
-
-        Builder(DslDataGenerator generator, File file) {
-            this.generator = generator;
-            this.file = file;
-            this.jsonString = null;
-            this.jsonNode = null;
-        }
-
-        Builder(DslDataGenerator generator, String jsonString) {
-            this.generator = generator;
-            this.file = null;
-            this.jsonString = jsonString;
-            this.jsonNode = null;
-        }
-
-        Builder(DslDataGenerator generator, JsonNode jsonNode) {
-            this.generator = generator;
-            this.file = null;
-            this.jsonString = null;
-            this.jsonNode = jsonNode;
-        }
-
-
-        /**
-         * Generates the data based on the configured DSL source.
-         *
-         * @return the generated data
-         * @throws IOException                                                        if file reading fails or JSON parsing fails
-         * @throws com.github.eddranca.datagenerator.exception.DslValidationException if DSL validation fails
-         */
-        public Generation generate() throws IOException {
-            if (file != null) {
-                return generator.generateInternal(file);
-            } else if (jsonString != null) {
-                return generator.generateInternal(jsonString);
-            } else if (jsonNode != null) {
-                return generator.generateInternal(jsonNode);
-            } else {
-                throw new IllegalStateException("No DSL source configured");
-            }
-        }
-
-        /**
-         * Generates the data and returns SQL INSERT statement streams for all collections.
-         *
-         * @return a map of table names to SQL INSERT statement streams
-         * @throws IOException if file reading fails
-         */
-        public Map<String, Stream<String>> generateAsSql() throws IOException {
-            return generate().asSqlInserts();
-        }
-
-        /**
-         * Generates the data and returns SQL INSERT statement streams for specified collections only.
-         *
-         * @param collectionNames the names of collections to generate SQL for
-         * @return a map of table names to SQL INSERT statement streams
-         * @throws IOException if file reading fails
-         */
-        public Map<String, Stream<String>> generateAsSql(String... collectionNames) throws IOException {
-            return generate().asSqlInserts(collectionNames);
-        }
-
-        /**
-         * Generates the data and returns JsonNode streams for all collections.
-         *
-         * @return a map of collection names to JsonNode streams
-         * @throws IOException if file reading fails
-         */
-        public Map<String, Stream<JsonNode>> generateAsJson() throws IOException {
-            return generate().asJsonNodes();
-        }
-
-        /**
-         * Generates the data and returns JsonNode streams for specified collections only.
-         *
-         * @param collectionNames the names of collections to generate JSON for
-         * @return a map of collection names to JsonNode streams
-         * @throws IOException if file reading fails
-         */
-        public Map<String, Stream<JsonNode>> generateAsJson(String... collectionNames) throws IOException {
-            return generate().asJsonNodes(collectionNames);
-        }
-
-        /**
-         * Generates the data and returns a JsonNode stream for a single collection.
-         *
-         * @param collectionName the name of the collection to stream
-         * @return a stream of JsonNode items
-         * @throws IOException              if file reading fails
-         * @throws IllegalArgumentException if the collection doesn't exist
-         */
-        public Stream<JsonNode> streamJsonNodes(String collectionName) throws IOException {
-            return generate().streamJsonNodes(collectionName);
-        }
-
-        /**
-         * Generates the data and returns a SQL INSERT stream for a single collection.
-         *
-         * @param collectionName the name of the collection to stream
-         * @return a stream of SQL INSERT statements
-         * @throws IOException              if file reading fails
-         * @throws IllegalArgumentException if the collection doesn't exist
-         */
-        public Stream<String> streamSqlInserts(String collectionName) throws IOException {
-            return generate().streamSqlInserts(collectionName);
-        }
-    }
 }
